jks/sun_crypto.py: Split off Sun JKS/JCE crypto functionality into jks/sun_crypto.py
