+++ jks/util.py
@@ -3,6 +3,12 @@ from __future__ import print_function
 import textwrap
 import base64
 
+RSA_ENCRYPTION_OID = (1,2,840,113549,1,1,1)
+DSA_OID            = (1,2,840,10040,4,1)       # identifier for DSA public/private keys; see RFC 3279, section 2.2.2 (e.g. in PKCS#8 PrivateKeyInfo or X.509 SubjectPublicKeyInfo)
+DSA_WITH_SHA1_OID  = (1,2,840,10040,4,3)       # identifier for the DSA signature algorithm; see RFC 3279, section 2.3.2 (e.g. in X.509 signatures)
+
+class KeystoreException(Exception): pass
+
 def as_hex(ba):
     return "".join("{:02x}".format(b) for b in bytearray(ba))
 
@@ -16,7 +22,7 @@ def print_pem(der_bytes, type):
     print(as_pem(der_bytes, type))
 
 def pkey_as_pem(pk):
-    if pk.algorithm_oid == (1,2,840,113549,1,1,1): # rsaEncryption
+    if pk.algorithm_oid == RSA_ENCRYPTION_OID:
         return as_pem(pk.pkey, "RSA PRIVATE KEY")
     else:
         return as_pem(pk.pkey_pkcs8, "PRIVATE KEY")
